pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-tavily>=0.2.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.21",
    "langchainhub>=0.1.21",
    "langgraph>=0.4.8",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.5",
]
[tool.pytest.ini_options]
testpaths = [
    "graph/chains/tests"
]
